% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-charts-prep.R
\name{data_charts_prep}
\alias{data_charts_prep}
\title{General function for processing general data}
\usage{
data_charts_prep(
  data,
  ftype,
  agg,
  color_by = NULL,
  ptage = FALSE,
  ptage_col = NULL,
  drop_na = FALSE,
  na_label = "na",
  drop_na_legend = TRUE,
  sort_opts = NULL,
  slice_n = NULL,
  palette = NULL,
  highlight_value = NULL,
  highlight_value_color = NULL,
  order_legend = NULL,
  order = NULL,
  label_wrap_legend = NULL,
  label_wrap = NULL,
  group_extra_num = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{ftype}{A string value with type of data to be plotted}

\item{agg}{Statistics which can be applied to all data subsets (sum, mean, median)}

\item{color_by}{A character with the name of the variable by which you want to color the graph. Default is NULL}

\item{ptage_col}{A string value with the name of the categorical variable against which the percentage is calculated.}

\item{group_extra_num}{A logical indicating}
}
\description{
General function for processing general data
}
\examples{

df <- sample_data("Cat-Num")
data_charts_prep(data = df, ftype = "Cat-Num". agg = "sum")


df <- sample_data("Cat-Cat-Num")
data_charts_prep(data = df, ftype = "Cat-Cat-Num", agg = "mean")

}
