% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-word-prep.R
\name{data_word_prep}
\alias{data_word_prep}
\title{Clean text to plot wordCloud}
\usage{
data_word_prep(data, ftype = NULL)
}
\arguments{
\item{data}{A data frame with geografical info.}

\item{ftype}{A string value with type of data to be plotted}
}
\description{
Clean text to plot wordCloud
}
