% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{format_prep}
\alias{format_prep}
\title{Format data}
\usage{
format_prep(data, dic, formats)
}
\arguments{
\item{data}{A data.frame}

\item{dic}{Dic of data to plot}

\item{formats}{list with sample_num, sample_cat, suffix and prefix}
}
\description{
function to format the strings or number of the data from scratch.
}
